service: Prevent respawn while stopping a service.
This is a followup to 5b08dcab1275e4b88fa0fba00aafb980c6ab4af7.

Contrary what to the comment says in that commit, it is possible for
SIGCHLD to be received and processed while the 'stop' method is running,
causing the service being stopped to be respawned.  To prevent this,
clear the 'running' slot before calling the service's 'stop' method so
that the process monitor doesn't end up respawning the thing.

* modules/shepherd/service.scm (stop): Set the 'running' slot to #f
before calling SERVICE's 'stop' method.  Remove inaccurate comment.
* tests/system-star.sh: Test that.