shepherd: Restore 'system*' in the child process after fork.
This is a followup to 975b0aa0d6f0a184f712c935cd1ad284cd39deaf.

Users such as the marionette service in Guix expect to be able to call
'primitive-fork' from shepherd and be able to run Guile code right from
there, without exec'ing.  This change makes sure they get a usable
'system*'.

* modules/shepherd.scm (main): Replace 'primitive-fork' with a procedure
that restores 'system*' in the child.