service: 'stop' blocks when a service is already being stopped.
This change allows the monitor to keep track of services that are being
stopped.  The 'stop' method now blocks when called on a service already
being stopped, until the service has actually been stopped.  This was
the case before 0.9.0; since 0.9.0 though, invoking 'stop' on a service
already being stopped would lead to a second invocation of its 'stop'
method.

* modules/shepherd/service.scm (stop): Send a 'stop' message
to (current-monitor-channel).  Call the 'stop' slot if and only if the
monitor returns a channel.
(service-monitor)[*service-stopped*, stopped-message?]: New variables.
Add 'stopping' loop variable and thread it in recursive calls.  Handle
'stop' and 'stopped-message?' messages.
* tests/stopping-status.sh: New file.
* Makefile.am (TESTS): Add it.