service: Controller and 'start' method inherit the right output port.
Previously, when creating and starting a service from the REPL (as shown
in the test), the controller and 'start' method would inherit
'current-output-port' from the REPL; that is, they would talk directly
over the REPL socket.  Once the REPL has terminated, the controller
would find itself attempting to write messages to a closed file port.
Likewise for the service's body if it inherited current-output-port from
its 'start' method.  This commit fixes that.

* modules/shepherd/comm.scm (%current-service-output-port): New
variable.
* modules/shepherd.scm (main): Parameterize it and parameterize
'current-output-port' accordingly.
* modules/shepherd/service.scm (spawn-service-controller): Parameterize
'current-output-port' and 'current-error-port'.
(start): Likewise around call to 'start'.
* tests/services/repl.sh: Add test.