service: One-shot services are started only once by 'start-in-parallel'.
Previously, since commit c195b87c275f39391c90a75a8843d4f6aedeb9c2, the
same one-shot service could be started several times if several services
depend on it.  This change ensures each one-shot service is started only
once.

* modules/shepherd/service.scm (%one-shot-services-started): New variable.
(start-in-parallel): Parameterize it.  Call 'lookup-services' on
elements of SERVICES that are symbols.  When starting a one-shot
service, call 'start' if and only if it's not already in
%ONE-SHOT-SERVICES-STARTED.
* tests/one-shot.sh: Add test.