service: Registry explicitly has only one service with a given name.
This was already the case before but parts of the code, such as the
'service-list' message handler in 'service-registry', maintained the
illusion that there could be several same-named services.

* modules/shepherd/service.scm (service-registry): Simplify
'service-list' based on the assumption that there can only be one
service for each name: registering a service with the same name as an
existing one turns it into a "replacement" for that service.
(lookup-canonical-service): Remove.
(fold-services): Expect a single service in the alist returned by
'service-list'.