service: Move service process monitoring to service controller.
This fixes a possible race condition whereby the 'find-service' call in
'process-monitor' could see the wrong running value, for instance
because the service is being started and hasn't yet recorded the PID as
its running value.

* modules/shepherd/service.scm (process-monitor): Remove call to
'find-service' and subsequent 'handle-service-termination' call.
(monitor-service-process): New procedure.
(handle-service-termination): Add PID parameter and pass it to the
controller.
(service-controller)[pid?]: New procedure.
In 'started-message?' and 'change-value-message?' clauses, add call to
'monitor-service-process' when NEW-VALUE looks like a PID.
Add 'pid' argument to 'handle-termination' message; leave value
unchanged when STATUS is 'stopped' or VALUE is not PID.