service: Removed unnecessary and racy 'handle-termination' messages.
* modules/shepherd/service.scm (service-controller): Do not send
'notify-termination' message after calling the 'handle-termination'
slot: the 'notify-termination' message could be received *before*
'handle-termination' has run and sent a 'handle-termination' message,
in which case we'd fail to respawn the service.
(stop): Do not send 'notify-termination' message: it is redundant with
the 'stopped-message?' sent once the 'stop' procedure has completed.