service: Wait when attempting to stop a service in "starting" state.
Previously, stopping a service would fail with "Service ~a might have
failed to stop." if it was in 'starting' state.  Now it will wait until
the service is running before attempting to stop it.

* modules/shepherd/service.scm (service-controller): In 'stop' clause,
when STATUS is 'starting, wait CONDITION then resend message to
CHANNEL.  Send #f on REPLY only if STATUS is 'stopped.
* tests/stopping-status.sh: Test it.